pyproject.toml
[tool.poetry]
name = "car-management-system"
version = "1.0.0"
description = "A Car Management System with SQLite database and PyQt6 GUI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
PyQt6 = "^6.4.0"
pyinstaller = "^6.11.1"
fpdf = "^1.7.2"
PyYAML = "^6.0"
python-dotenv = "^1.0.0"
cx-freeze = "^7.2.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-qt = "^4.2.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pyinstaller = "^6.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
car-manager = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"